requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
Pillow==6.2.1
pyocr==0.7.2
requests==2.22.0
urllib3==1.25.6
